fix(geometry3d): move Point3 into SMath.Geometry3D namespace
Point3.cs lived in the Geometry3D folder but declared the
SMath.Geometry2D namespace, the sole file in the library whose
namespace did not match its category folder. It is a 3D type
((X, Y, Z) tuples) and its folder-sibling Sphere.cs already used
SMath.Geometry3D, so this was a copy-paste artifact from Point2.cs.

Move it to SMath.Geometry3D and align the test files (Point3Tests,
SphereTests) to the same namespace so the Geometry3D test folder is
consistent.
