+++ src/quality/SMath__Tests/Geometry3D/Point3Tests.cs
@@ -1,7 +1,7 @@
-﻿using SMath.Geometry2D;
+﻿using SMath.Geometry3D;
 using Xunit;
 
-namespace SMath.Geometry2D;
+namespace SMath.Geometry3D;
 
 public class Point3Tests
 {
